requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
langchain
openai
crewai
duckduckgo-search
pydantic
requests
phidata
nest_asyncio